fix campos - @vinilope

+++ src/main/java/com/serenity/api/serenity/dtos/escala/EscalaResponse.java
@@ -8,7 +8,7 @@ public record EscalaResponse(
         String funcaoEscala,
         Integer qtdColaborador,
         Integer qtdHora,
-        Integer valor,
+        Double valor,
         Boolean comissionado,
         Boolean asoObrigatorio
 ) {
